wip, added back poetry
