+++ .gitignore
@@ -135,8 +135,6 @@ cache
 *.zip
 cache.json
 Untitled*.ipynb
-pyproject.toml
-poetry.lock
 tmp/
 lab/
 *.csv
