+++ icortex/__init__.py
@@ -1,4 +1,3 @@
 from icortex.kernel import ICortexKernel, print_help, get_icortex_kernel
-from icortex.cli import set_icortex_service, eval_cli
-from icortex.exec import eval_prompt
+# from icortex.cli import set_icortex_service, eval_cli
 import icortex.services
