+++ pyproject.toml
@@ -1,0 +1,27 @@
+[tool.poetry]
+name = "icortex"
+version = "0.0.1"
+description = "Jupyter kernel that allows you generate code from natural language prompts"
+authors = ["TextCortex Team <onur@textcortex.com>"]
+license = "Apache"
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.10"
+click = "^8.1.3"
+entrypoints = "^0.4"
+ipykernel = "^6.16.0"
+ipython = "^8.5.0"
+ipywidgets = "^8.0.2"
+jupyter-client = "^7.4.2"
+jupyter-console = "^6.4.4"
+jupyter-core = "^4.11.1"
+jupyterlab-widgets = "^3.0.3"
+openai = "^0.23.1"
+Pygments = "^2.13.0"
+toml = "^0.10.2"
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
